pyproject.toml
[tool.poetry]
name = "xp-pen"
version = "0.1.0"
description = ""
authors = ["Todd Sifleet <todd.sifleet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyusb = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"